Updated tests and remove RecoveryStatus, DefaultRecoveryScheduler and RecoveryResource
