+++ src/main/java/org/apache/mesos/scheduler/recovery/monitor/TestingFailureMonitor.java
@@ -1,6 +1,7 @@
 package org.apache.mesos.scheduler.recovery.monitor;
 
 import org.apache.mesos.Protos.TaskInfo;
+import org.apache.mesos.scheduler.recovery.FailureUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,6 +26,7 @@ public class TestingFailureMonitor implements FailureMonitor {
         failedList = new ArrayList<>();
         for (TaskInfo info : failed) {
             failedList.add(info);
+            FailureUtils.markFailed(info);
         }
     }
